requirements.txt
fpdf2
fastapi
jupyter
matplotlib
numpy
pandas
pyyaml
uvicorn
